+++ libs/esp32/pxt.json
@@ -4,7 +4,6 @@
     "files": [
         "esp32spi.ts",
         "socket.ts",
-        "requests.ts",
         "net.ts",
         "README.md"
     ],
